Revert "sys/src/9/rb: import rb changes (2020-12-21) (thanks Geoff Collyer)"
This reverts commit 1eb0e3e3ae0d4da969cc1ece4f8b4b649cd64add.

+++ sys/src/9/rb/faultmips.c
@@ -226,10 +226,10 @@ faultmips(Ureg *ur, int user, int code)
 }
 
 /*
- * called in sysfile.c
+ * called in syscallfmt.c, sysfile.c, sysproc.c
  */
 void
-validalign(ulong addr, ulong align)
+validalign(uintptr addr, unsigned align)
 {
 	/*
 	 * Plan 9 is a 32-bit O/S, and the hardware it runs on
